requirements.txt
BeautifulSoup==3.2.1
Django==1.5.4
dulwich>=0.9.4
Markdown==2.3.1
Pygments==1.6
Whoosh==2.5.6
argparse>=1.1
async==0.6.1
django-haystack==2.1.0
gitdb==0.5.4
pytz==2011n
repoze.timeago==0.5
smmap>=0.8.2
wsgiref==0.1.2
docutils==0.11
python-social-auth>=0.1.20
requests>=2.2.1
oauthlib>=0.6.1
six>=1.5.2
requests-oauthlib>=0.4.0